Add prefix framework
To add the first set of macros which call AC_DEFINE or AC_SUBST,
we need a way of setting a macro prefix, so that we do not
need to do drastic overhauls of calling projects and their source
code.

Rename oac_literal.m4 to oac_assert.m4 and add a macro to enforce
evaluation ordering.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>