No docs without a properly installed sphinx (with modules).
The logic to install the dpc was missing a protection around the case
where sphinx was installed but the modules required for OMPI were not. I
slightly change the order of operations to ensure that we never try to
install the docs if there were not available (aka not from a tarball) or
not properly generated.

Signed-off-by: George Bosilca <gbosilca@nvidia.com>